service_mcastproxy.sh: only start mcproxy_v4 when erouter mode is "ipv4Only" or "ipv4AndIpv6" ( OFW-2446 )
Add data model support for Device.X_LGI-COM_Multicast.IGMPv3ProxyEnable ( OFW-2446 )

[Issue]
Update service_mcastproxy.sh according to MVXREQ-1814 and MVXREQ-1818

[Solution]
- start mcproxy_v4 only if lgi_igmpproxy_enable is 1 and the router is not in DSLite or
  ipv6Only mode

[Testing Performed]
1. Enable DSLite and verify that mcproxy_v4 is not running
2. Disable DSLite and verify that mcproxy_v4 is running
3. verify that mcproxy_v4 is running when erouter mode is "ipv4Only" or "ipv4AndIpv6"

+++ source/scripts/init/service.d/service_mcastproxy.sh
@@ -94,19 +94,21 @@ fi
    fi
 
 else
-   cat $LOCAL_CONF_FILE > $CONF_FILE
-   rm -f $LOCAL_CONF_FILE
-   if [ "$BOX_TYPE" = "HUB4" ] || [ "$BOX_TYPE" = "SR300" ]; then
-       $BIN $CONF_FILE &
-   else 
-       $BIN -r -f $CONF_FILE &
+   if [ "$SYSCFG_last_erouter_mode" = "1" ] || [ "$SYSCFG_last_erouter_mode" = "3" ]; then
+      cat $LOCAL_CONF_FILE > $CONF_FILE
+      rm -f $LOCAL_CONF_FILE
+      if [ "$BOX_TYPE" = "HUB4" ] || [ "$BOX_TYPE" = "SR300" ]; then
+         $BIN $CONF_FILE &
+      else
+         $BIN -r -f $CONF_FILE &
+      fi
    fi
 fi
 }
 
 service_init ()
 {
-   queries="igmpproxy_enabled lan_ifname dslite_enable dslite_tunnel_interface_1"
+   queries="igmpproxy_enabled lan_ifname dslite_enable dslite_tunnel_interface_1 last_erouter_mode"
    get_utctx_val "$queries"
    eval `sysevent batchget current_wan_ifname wan-status lan-status bridge-status`
    WAN_IFNAME=$SYSEVENT_1
